backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv

langchain
langchain-text-splitters
langchain-community
langchain-huggingface
chromadb
pypdf
transformers
sentencepiece
sentence-transformers